Revert "Add support for mounts on Darwin"
This reverts commit 2799b28e61f99336b776ef537bc7e2ab2fcffc3c.
